MatrixPilot_fbw - Trunk merge to branch done and part tested
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_fbw@2327 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/gain_variables.h
@@ -11,10 +11,10 @@
 // YAWKD_AILERON is the derivative feedback gain for ailerons in response to yaw rotation
 // AILERON_BOOST is the additional gain multiplier for the manually commanded aileron deflection
 
-extern int rollkp;
-extern int rollkd;
-extern int yawkpail;
-extern int yawkdail;
+extern int16_t rollkp;
+extern int16_t rollkd;
+extern int16_t yawkpail;
+extern int16_t yawkdail;
 
 //#define AILERON_BOOST						1.0
 
@@ -24,10 +24,10 @@ extern int yawkdail;
 // RUDDER_ELEV_MIX is the degree of elevator adjustment for rudder and banking
 // AILERON_ELEV_MIX is the degree of elevator adjustment for aileron
 // ELEVATOR_BOOST is the additional gain multiplier for the manually commanded elevator deflection
-extern int pitchgain;
-extern int pitchkd;
-extern int rudderElevMixGain;
-extern int rollElevMixGain;
+extern int16_t pitchgain;
+extern int16_t pitchkd;
+extern int16_t rudderElevMixGain;
+extern int16_t rollElevMixGain;
 //#define ELEVATOR_BOOST						0.5
 
 // Neutral pitch angle of the plane (in degrees) when flying inverted
@@ -42,11 +42,11 @@ extern int rollElevMixGain;
 // MANUAL_AILERON_RUDDER_MIX is the fraction of manual aileron control to mix into the rudder when
 // in stabilized or waypoint mode.  This mainly helps aileron-initiated turning while in stabilized.
 // RUDDER_BOOST is the additional gain multiplier for the manually commanded rudder deflection
-extern int yawkprud;
-extern int yawkdrud;
-extern int rollkprud;
-extern int rollkdrud;
-//extern int MANUAL_AILERON_RUDDER_MIX			0.20
+extern int16_t yawkprud;
+extern int16_t yawkdrud;
+extern int16_t rollkprud;
+extern int16_t rollkdrud;
+//extern int16_t MANUAL_AILERON_RUDDER_MIX			0.20
 //#define RUDDER_BOOST						1.0
 
 // Gains for Hovering
@@ -63,12 +63,12 @@ extern int rollkdrud;
 // HOVER_NAV_MAX_PITCH_RADIUS is the radius around a waypoint in meters, within which the HOVER_PITCH_TOWARDS_WP
 //                            value is proportionally scaled down.
 
-extern int hoverrollkp;
-extern int hoverrollkd;
-extern int hoverpitchgain;
-extern int hoverpitchkd;
-extern int hoveryawkp;
-extern int hoveryawkd;
+extern int16_t hoverrollkp;
+extern int16_t hoverrollkd;
+extern int16_t hoverpitchgain;
+extern int16_t hoverpitchkd;
+extern int16_t hoveryawkp;
+extern int16_t hoveryawkd;
 
 //#define HOVER_PITCH_OFFSET					0.0		// + leans towards top, - leans towards bottom
 //#define HOVER_YAW_OFFSET					0.0
@@ -78,22 +78,22 @@ extern int hoveryawkd;
 
 /*
 // servo throw can be more than 3 turns - 1080 degrees - so use integers rather than char
-const int tan_pitch_in_stabilized_mode;
-const int yaw_in_stabilized_mode;
+const int16_t tan_pitch_in_stabilized_mode;
+const int16_t yaw_in_stabilized_mode;
 
-const int pitch_offset_centred;
-const int yaw_offset_centred;
+const int16_t pitch_offset_centred;
+const int16_t yaw_offset_centred;
 
-const int pitch_servo_max;
-const int pitch_servo_min;
-const int yaw_servo_max;
-const int yaw_servo_min;
+const int16_t pitch_servo_max;
+const int16_t pitch_servo_min;
+const int16_t yaw_servo_max;
+const int16_t yaw_servo_min;
 
 // servo_ratios are used to convert degrees of rotation into servo pulse code lengths
 // This code is configured for the full throw of the servo to be achieved by a range of
 // 2000 units being sent to udb_pwOut. (i.e. min 2000, centered 3000, max 4000 )
-const int pitch_servo_ratio;
-const int yaw_servo_ratio;
+const int16_t pitch_servo_ratio;
+const int16_t yaw_servo_ratio;
 
 */
